+++ src/map/elemental.cpp
@@ -5,7 +5,6 @@
 #include "../common/malloc.h"
 #include "../common/timer.h"
 #include "../common/nullpo.h"
-#include "../common/mmo.h"
 #include "../common/showmsg.h"
 #include "../common/random.h"
 #include "../common/strlib.h"
